[fix] Move workspace role assign/unassign endpoints to OSS
The RBAC migration moved enforcement to OSS but left the role
assign/unassign endpoints, update_user_roles, and the UserRole model in
EE — so changing a member's role in OSS hit a 404 on
POST /workspaces/{id}/roles.

- Add POST/DELETE /workspaces/{id}/roles/ to the OSS workspace router,
  gated by check_action_access(MODIFY_USER_ROLES).
- Port update_user_roles into OSS db_manager (owner-guard, project-member
  sync, cache invalidation preserved); add UserRole to OSS workspace_models.
- Drop the now-duplicate EE handlers and update_user_roles; EE re-exports
  the OSS function (avoids route collision when EE mounts over OSS).
- Gate the Members "Remove"/"Resend invitation" actions on canModifyRoles;
  drop the redundant (!isEE() || hasRBAC) on the role-edit dropdown.
