+++ api/oss/src/services/db_manager.py
@@ -50,6 +50,7 @@ from oss.src.models.db_models import (
     ProjectMemberDB,
 )
 from oss.src.dbs.postgres.webhooks.dbes import WebhookSubscriptionDBE
+from oss.src.models.api.workspace_models import UserRole
 from oss.src.core.testcases.dtos import Testcase
 from oss.src.core.testsets.dtos import (
     TestsetRevisionData,
@@ -1552,6 +1553,110 @@ async def sync_workspace_members_to_project(
     engine = get_transactions_engine()
     async with engine.session() as new_session:
         await _sync(new_session)
+
+
+async def update_user_roles(
+    workspace_id: str,
+    payload: UserRole,
+    delete: bool = False,
+) -> bool:
+    """
+    Update a user's roles in a workspace (and mirror onto its projects).
+
+    Args:
+        workspace_id (str): The ID of the workspace.
+        payload (UserRole): The user email and role to update.
+        delete (bool): Whether to clear the role assignment instead of setting it.
+
+    Returns:
+        bool: True if the user's roles were successfully updated.
+    """
+
+    user = await get_user_with_email(payload.email)
+    projects = await fetch_projects_by_workspace(workspace_id)
+    if not projects:
+        raise NoResultFound(
+            f"No projects found for the provided workspace_id {workspace_id}"
+        )
+
+    engine = get_transactions_engine()
+
+    async with engine.session() as session:
+        workspace_member_result = await session.execute(
+            select(WorkspaceMemberDB).filter_by(
+                workspace_id=uuid.UUID(workspace_id), user_id=user.id
+            )
+        )
+        workspace_member = workspace_member_result.scalars().first()
+        if not workspace_member:
+            raise NoResultFound(
+                f"User with id {str(user.id)} is not part of the workspace member."
+            )
+
+        if workspace_member.role == "owner":
+            raise HTTPException(
+                403,
+                {
+                    "message": "You do not have permission to perform this action. Please contact your Organization Owner"
+                },
+            )
+
+        project_ids = [project.id for project in projects]
+        project_members_result = await session.execute(
+            select(ProjectMemberDB).filter(
+                ProjectMemberDB.project_id.in_(project_ids),
+                ProjectMemberDB.user_id == user.id,
+            )
+        )
+        project_members = project_members_result.scalars().all()
+        if len(project_members) != len(project_ids):
+            for project in projects:
+                await sync_workspace_members_to_project(
+                    str(project.id), session=session
+                )
+
+            project_members_result = await session.execute(
+                select(ProjectMemberDB).filter(
+                    ProjectMemberDB.project_id.in_(project_ids),
+                    ProjectMemberDB.user_id == user.id,
+                )
+            )
+            project_members = project_members_result.scalars().all()
+
+        if len(project_members) != len(project_ids):
+            raise NoResultFound(
+                f"User with id {str(user.id)} is not part of all project memberships."
+            )
+
+        if not delete:
+            workspace_member.role = payload.role
+            for member in project_members:
+                member.role = payload.role
+
+        await session.commit()
+
+        default_project_id = next(
+            (project.id for project in projects if project.is_default),
+            projects[0].id,
+        )
+        default_project_member = next(
+            (
+                member
+                for member in project_members
+                if member.project_id == default_project_id
+            ),
+            None,
+        )
+        if default_project_member:
+            await session.refresh(default_project_member)
+
+    for project in projects:
+        await invalidate_cache(
+            namespace="check_action_access",
+            project_id=str(project.id),
+        )
+
+    return True
 
 
 async def delete_project(project_id: str) -> None:
